+++ openrouteservice-api-tests/src/test/java/org/heigit/ors/v2/services/isochrones/ResultTest.java
@@ -42,7 +42,7 @@ public class ResultTest extends ServiceTest {
 
         JSONArray secondLocation = new JSONArray();
         secondLocation.put(8.684177);
-        secondLocation.put(49.411034);
+        secondLocation.put(49.410034);
 
         JSONArray unknownLocation = new JSONArray();
         unknownLocation.put(-18.215332);
@@ -118,9 +118,9 @@ public class ResultTest extends ServiceTest {
                 .then()
                 .body("any { it.key == 'type' }", is(true))
                 .body("any { it.key == 'features' }", is(true))
-                .body("features[0].geometry.coordinates[0].size()", is(52))
+                .body("features[0].geometry.coordinates[0].size()", is(49))
                 .body("features[0].properties.center.size()", is(2))
-                .body("bbox", hasItems(8.663811f, 49.409103f, 8.699429f, 49.43929f))
+                .body("bbox", hasItems(8.663323f, 49.40837f, 8.700336f, 49.439884f))
                 .body("features[0].type", is("Feature"))
                 .body("features[0].geometry.type", is("Polygon"))
                 .body("features[0].properties.group_index", is(0))
@@ -189,10 +189,10 @@ public class ResultTest extends ServiceTest {
                 .when()
                 .post(getEndPointPath() + "/{profile}/geojson")
                 .then()
-                .body("bbox[0]", is(8.663811f))
-                .body("bbox[1]", is(49.409103f))
-                .body("bbox[2]", is(8.699429f))
-                .body("bbox[3]", is(49.439289f))
+                .body("bbox[0]", is(8.663323f))
+                .body("bbox[1]", is(49.40837f))
+                .body("bbox[2]", is(8.700336f))
+                .body("bbox[3]", is(49.439884f))
                 .statusCode(200);
     }
 
@@ -214,8 +214,8 @@ public class ResultTest extends ServiceTest {
                 .then()
                 .body("any { it.key == 'type' }", is(true))
                 .body("any { it.key == 'features' }", is(true))
-                .body("features[0].properties.area", is(both(greaterThan(6110000f)).and(lessThan(6120000f))))
-                .body("features[0].properties.reachfactor", is(0.7007f))
+                .body("features[0].properties.area", is(both(greaterThan(6590000f)).and(lessThan(6600000f))))
+                .body("features[0].properties.reachfactor", is(0.7561f))
                 .statusCode(200);
 
     }
@@ -239,8 +239,8 @@ public class ResultTest extends ServiceTest {
                 .then()
                 .body("any { it.key == 'type' }", is(true))
                 .body("any { it.key == 'features' }", is(true))
-                .body("features[0].properties.area", is(both(greaterThan(6110000f)).and(lessThan(6120000f))))
-                .body("features[0].properties.reachfactor", is(0.7007f))
+                .body("features[0].properties.area", is(both(greaterThan(6590000f)).and(lessThan(6600000f))))
+                .body("features[0].properties.reachfactor", is(0.7561f))
                 .statusCode(200);
 
     }
@@ -264,8 +264,8 @@ public class ResultTest extends ServiceTest {
                 .then()
                 .body("any { it.key == 'type' }", is(true))
                 .body("any { it.key == 'features' }", is(true))
-                .body("features[0].properties.area", is(both(greaterThan(6.11f)).and(lessThan(6.12f))))
-                .body("features[0].properties.reachfactor", is(0.7007f))
+                .body("features[0].properties.area", is(both(greaterThan(6.59f)).and(lessThan(6.60f))))
+                .body("features[0].properties.reachfactor", is(0.7561f))
                 .statusCode(200);
 
     }
@@ -291,7 +291,7 @@ public class ResultTest extends ServiceTest {
                 .then()
                 .body("any { it.key == 'type' }", is(true))
                 .body("any { it.key == 'features' }", is(true))
-                .body("features[0].properties.area", is(both(greaterThan(6.11f)).and(lessThan(6.12f))))
+                .body("features[0].properties.area", is(both(greaterThan(6.59f)).and(lessThan(6.60f))))
                 .statusCode(200);
 
     }
@@ -315,8 +315,8 @@ public class ResultTest extends ServiceTest {
                 .then()
                 .body("any { it.key == 'type' }", is(true))
                 .body("any { it.key == 'features' }", is(true))
-                .body("features[0].properties.area", is(both(greaterThan(2.36f)).and(lessThan(2.37f))))
-                .body("features[0].properties.reachfactor", is(0.7007f))
+                .body("features[0].properties.area", is(both(greaterThan(2.53f)).and(lessThan(2.55f))))
+                .body("features[0].properties.reachfactor", is(0.7561f))
                 .statusCode(200);
 
     }
@@ -348,11 +348,11 @@ public class ResultTest extends ServiceTest {
                 .body("features[2].type", is("Feature"))
                 .body("features[2].geometry.type", is("Polygon"))
                 //.body("features[2].geometry.coordinates[0].size()", is(26))
-                .body("features[2].geometry.coordinates[0].size()", is(35))
+                .body("features[2].geometry.coordinates[0].size()", is(38))
                 .body("features[2].properties.contours.size()", is(2))
                 .body("features[2].properties.containsKey('area')", is(true))
                 //.body("features[2].properties.area", is(5824280.5f))
-                .body("features[0].properties.area", is(both(greaterThan(6110000f)).and(lessThan(6120000f))))
+                .body("features[0].properties.area", is(both(greaterThan(6590000f)).and(lessThan(6600000f))))
                 .body("features[2].properties.contours[0][0]", is(0))
                 .body("features[2].properties.contours[0][1]", is(0))
                 .body("features[2].properties.contours[1][0]", is(1))