+++ CHANGELOG.md
@@ -37,6 +37,9 @@ RELEASING:
 ### Added
 - API endpoint "centrality" to calculate [betweenness centrality](https://en.wikipedia.org/wiki/Betweenness_centrality) values for nodes inside a given bounding box. Centrality is calculated using Brandes' algorithm. 
 - Support for turn restrictions with core-based routing algorithms
+### Changed
+- Use Querygraph and virtual nodes for isochrone calculation in the same manner as in routing
+- Remove Isochrones v1 api tests
 ### Fixed
 - Fixed calculation of route distance limits with skipped segments ([#814](https://github.com/GIScience/openrouteservice/issues/814))
 - fixed missing segment distance and duration ([#695](https://github.com/GIScience/openrouteservice/issues/695)